Feature/#14 comment reaction

+++ src/main/java/com/kgat/repository/ArticleRepository.java
@@ -18,7 +18,7 @@ public interface ArticleRepository extends JpaRepository<Article, Integer> {
     Page<Article> findByArticleWriter(String articleWriter, Pageable pageable);
 
     // 작성자명과 게시글 번호로 삭제
-    void deleteByArticleWriterAndArticleNum(String userId, Integer articleNum);
+    void deleteByArticleWriterAndArticleNum(String userId, Long articleNum);
 
     Article findByArticleNum(Long articleNum);
 }
